+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
             android:usesCleartextTraffic="true"
             tools:targetApi="31">
         <activity
-                android:name=".MainActivity"
+                android:name=".view.MainActivity"
                 android:exported="true"
                 android:theme="@style/SplashTheme"
                 android:launchMode="singleTask">
@@ -64,7 +64,7 @@
             </intent-filter>
         </activity> <!-- 서비스를 추가하고 인텐트 필터를 설정한다. -->
         <service
-                android:name=".FcmAppService"
+                android:name=".service.FcmAppService"
                 android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT"/>
