+++ app/src/main/java/com/hmoa/app/service/FcmAppService.kt
@@ -1,4 +1,4 @@
-package com.hmoa.app
+package com.hmoa.app.service
 
 import android.app.Notification
 import android.app.NotificationChannel
@@ -8,28 +8,36 @@ import android.content.Context
 import android.content.Intent
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
+import com.hmoa.app.view.MainActivity
 
 class FcmAppService : FirebaseMessagingService() {
     override fun onNewToken(token: String) {
         super.onNewToken(token)
     }
+
     override fun onMessageReceived(remoteMessage: RemoteMessage) {
         super.onMessageReceived(remoteMessage)
 
-        if(remoteMessage.data.isNotEmpty()){
+        if (remoteMessage.data.isNotEmpty()) {
             sendNotification(remoteMessage.data)
         }
     }
-    private fun sendNotification(fcmData : Map<String, String>){
+
+    private fun sendNotification(fcmData: Map<String, String>) {
         val CHANNEL_DEFAULT_IMPORTANCE = "channelId"
         val ONGOING_NOTIFICATION = 1
 
-        val notificationIntent = Intent(this, MainActivity::class.java).apply{
+        val notificationIntent = Intent(this, MainActivity::class.java).apply {
             addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
-            putExtra("deeplink",fcmData["deeplink"])
-            putExtra("id",fcmData["id"])
+            putExtra("deeplink", fcmData["deeplink"])
+            putExtra("id", fcmData["id"])
         }
-        val pendingIntent = PendingIntent.getActivity(this, 0, notificationIntent, PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT)
+        val pendingIntent = PendingIntent.getActivity(
+            this,
+            0,
+            notificationIntent,
+            PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
+        )
 
         val notification = Notification.Builder(this, CHANNEL_DEFAULT_IMPORTANCE)
             .setContentTitle(fcmData["title"])
@@ -41,7 +49,8 @@ class FcmAppService : FirebaseMessagingService() {
 
         val notificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
 
-        val channel = NotificationChannel(CHANNEL_DEFAULT_IMPORTANCE, "HMOA Channel", NotificationManager.IMPORTANCE_DEFAULT)
+        val channel =
+            NotificationChannel(CHANNEL_DEFAULT_IMPORTANCE, "HMOA Channel", NotificationManager.IMPORTANCE_DEFAULT)
         notificationManager.createNotificationChannel(channel)
 
         notificationManager.notify(ONGOING_NOTIFICATION, notification)
