+++ app/src/main/java/com/hmoa/app/viewmodel/AppViewModel.kt
@@ -1,4 +1,4 @@
-package com.hmoa.app
+package com.hmoa.app.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -7,8 +7,6 @@ import com.hmoa.core_domain.repository.LoginRepository
 import com.hmoa.core_model.request.FCMTokenSaveRequestDto
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
